Allow user to add new books
Added a form for the users to input a title, author, number of pages, and read status, and an event listener that creates a new entry into myLibrary and adds the user's input. There is also now a displayNewBook function that adds the new book to the table for display, separately from the displayLibrary function.