Fix input for having already read book
The input element of the form where the user selects whether they have read the book has been fixed. Previously, it relied on the value attribute and would not update based on the user input. It now has an (inputRead.checked == true) conditional to ensure the new book's read property is always up-to-date with the user input.