Add index class to all table entries
Adding a book to the library now adds a class of 'index[number]' to its <tr> and all its child <td> elements (as in, the first book's row and boxes all have the 'index1' class, the second book's all have the 'index2' class, and so on. This will enable a specific book's row to be deleted by that book's associated Delete Book button.